[602] Move listening mode configuration to separate file
